+++ src/aux-server/aux-backend/prisma/PrismaFileRecordsLookup.ts
@@ -235,6 +235,7 @@ export class PrismaFileRecordsLookup implements FileRecordsLookup {
                         recordName,
                         fileName,
                     },
+                    uploadedAt: { equals: null },
                 },
                 data: {
                     uploadedAt: new Date(),