+++ CHANGELOG.md
@@ -15,6 +15,8 @@
         -   `absolute` - The mesh won't be repositioned. It will retain the position configured in the GLTF.
 -   Added support for [Anthropic AI Chat models](https://docs.anthropic.com/en/docs/about-claude/models) to `ai.chat()` and `ai.stream.chat()`.
     -   `SERVER_CONFIG` needs to be configured with an Anthropic API Key and allowed `anthropic` models.
+-   Add support for [Minio](https://min.io/) for file records.
+    -   Can be configured by using the `minio` property in `SERVER_CONFIG`.
 
 ### :bug: Bug Fixes