+++ CHANGELOG.md
@@ -15,6 +15,7 @@
         -   `absolute` - The mesh won't be repositioned. It will retain the position configured in the GLTF.
 -   Added support for [Anthropic AI Chat models](https://docs.anthropic.com/en/docs/about-claude/models) to `ai.chat()` and `ai.stream.chat()`.
     -   `SERVER_CONFIG` needs to be configured with an Anthropic API Key and allowed `anthropic` models.
+-   Added support for [`stable-image-ultra`](https://platform.stability.ai/docs/api-reference#tag/Generate/paths/~1v2beta~1stable-image~1generate~1ultra/post), [`stable-image-core`](https://platform.stability.ai/docs/api-reference#tag/Generate/paths/~1v2beta~1stable-image~1generate~1core/post), [`sd3-medium`, `sd3-large`, `sd3-large-turbo`](https://platform.stability.ai/docs/api-reference#tag/Generate/paths/~1v2beta~1stable-image~1generate~1sd3/post) AI image generation models.
 -   Add support for [Minio](https://min.io/) for file records.
     -   Can be configured by using the `minio` property in `SERVER_CONFIG`.